Entity @Setter 어노테이션 삭제, builder 사용 및 Service interface 삭제 후 병합
