+++ src/main/java/com/quostomize/quostomize_be/domain/customizer/pointUsageMethod/Service/PointUsageMethodService.java
@@ -1,12 +1,67 @@
 package com.quostomize.quostomize_be.domain.customizer.pointUsageMethod.Service;
 
+import com.quostomize.quostomize_be.common.error.ErrorCode;
+import com.quostomize.quostomize_be.common.error.exception.AppException;
 import com.quostomize.quostomize_be.domain.customizer.pointUsageMethod.entity.PointUsageMethod;
+import com.quostomize.quostomize_be.domain.customizer.pointUsageMethod.repository.PointUsageMethodRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-public interface PointUsageMethodService {
-    // usageType : 포인트 사용방법의 타입(Lotto, Payback, PieceStock)
 
-    PointUsageMethod getPointUsageMethod(Long cardSequenceId);
+@Service
+@RequiredArgsConstructor
+public class PointUsageMethodService {
 
-    PointUsageMethod togglePointUsage(Long cardSequenceId, String usageType, boolean isActive);
+    private final PointUsageMethodRepository repository;
+
+    PointUsageMethod pointUsageMethod = PointUsageMethod.builder()
+            .isLotto(true)
+            .isPayback(false)
+            .isPieceStock(true)
+            .build();
+
+
+    public PointUsageMethod getPointUsageMethod(Long cardSequenceId) {
+
+        return repository.findByCardDetail_CardSequenceId(cardSequenceId)
+                .orElseThrow(()-> new AppException(ErrorCode.CARD_NOT_FOUND));
+    }
+
+
+    public PointUsageMethod togglePointUsage(Long cardSequenceId, String usageType, boolean isActive) {
+        PointUsageMethod pointUsageMethod = repository.findByCardDetail_CardSequenceId(cardSequenceId)
+                .orElseThrow(() -> new AppException(ErrorCode.CARD_NOT_FOUND));
+
+
+        PointUsageMethod updatedPointUsage;
+
+        // 사용 유형에 따른 필드 변경
+        switch (usageType) {
+            case "lotto":
+                updatedPointUsage = PointUsageMethod.builder()
+                                                    .isLotto(isActive)
+                                                    .build();
+                break;
+            case "payback":
+                if (isActive && pointUsageMethod.getIsPieceStock()) {
+                    throw new AppException(ErrorCode.PAYBACK_AND_PIECESTOCK_CONFLICT);
+                }
+                updatedPointUsage = PointUsageMethod.builder()
+                                                    .isPayback(isActive)
+                                                    .build();
+                break;
+            case "pieceStock":
+                if (isActive && pointUsageMethod.getIsPayback()) {
+                    throw new AppException(ErrorCode.PAYBACK_AND_PIECESTOCK_CONFLICT);
+                }
+                updatedPointUsage = PointUsageMethod.builder()
+                                                    .isPieceStock(isActive)
+                                                    .build();
+                break;
+        }
+
+        return repository.save(pointUsageMethod);
+    }
+
+
 }
