+++ src/main/java/com/quostomize/quostomize_be/domain/customizer/pointUsageMethod/entity/PointUsageMethod.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.quostomize.quostomize_be.common.entity.BaseTimeEntity;
 import com.quostomize.quostomize_be.domain.customizer.card.entity.CardDetail;
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -33,12 +34,12 @@ public class PointUsageMethod extends BaseTimeEntity {
     @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private CardDetail cardDetail;
 
-    public void setIsLotto(boolean isActive) {
+    @Builder
+    public PointUsageMethod(Boolean isLotto, Boolean isPayback, Boolean isPieceStock, CardDetail cardDetail) {
+        this.isLotto = isLotto;
+        this.isPayback = isPayback;
+        this.isPieceStock = isPieceStock;
+        this.cardDetail = cardDetail;
     }
 
-    public void setIsPayback(boolean isActive) {
-    }
-
-    public void setIsPieceStock(boolean isActive) {
-    }
 }
